+++ pyodide-build/pyodide_build/pywasmcross.py
@@ -10,7 +10,7 @@ The gist is:
   wrappers that store the arguments in a log, and then delegate along to the
   real native compiler and linker.
 
-- Remove all of the native build products.
+- Remove all the native build products.
 
 - Play back the log, replacing the native compiler with emscripten and
   adjusting include paths and flags as necessary for cross-compiling to
@@ -109,7 +109,7 @@ def collect_args(basename):
 
 def make_symlinks(env):
     """
-    Makes sure all of the symlinks that make this script look like a compiler
+    Makes sure all the symlinks that make this script look like a compiler
     exist.
     """
     TOOLSDIR = Path(common.get_make_flag("TOOLSDIR"))
@@ -313,11 +313,19 @@ def handle_command(line, args, dryrun=False):
         # some gcc flags that clang does not support actually
         if arg == "-Bsymbolic-functions":
             continue
-        if arg == "-Wl,-Bsymbolic-functions":
-            continue
-        # breaks emscripten see https://github.com/emscripten-core/emscripten/issues/14460
-        if arg == "-Wl,--strip-all":
-            continue
+        # ignore some link flags
+        # it should not check if `arg == "-Wl,-xxx"` and ignore directly here,
+        # because arg may be something like "-Wl,-xxx,-yyy" where we only want
+        # to ignore "-xxx" but not "-yyy".
+        if arg.startswith("-Wl"):
+            arg = arg.replace(",-Bsymbolic-functions", "")
+            # breaks emscripten see https://github.com/emscripten-core/emscripten/issues/14460
+            arg = arg.replace(",--strip-all", "")
+            # wasm-ld does not regconize some link flags
+            arg = arg.replace(",--sort-common", "")
+            arg = arg.replace(",--as-needed", "")
+            if arg == "-Wl":
+                continue
         # threading is disabled for now
         if arg == "-pthread":
             continue
@@ -328,6 +336,9 @@ def handle_command(line, args, dryrun=False):
         if arg in ["-bundle", "-undefined", "dynamic_lookup"]:
             continue
         if arg == "-lffi":
+            continue
+        # This flag is needed to build numpy with SIMD optimization
+        if arg == "-mpopcnt":
             continue
         # The native build is possibly multithreaded, but the emscripten one
         # definitely isn't