+++ src/core/hiwire.h
@@ -4,6 +4,7 @@
 #include "Python.h"
 #include "stdalign.h"
 #include "types.h"
+#define WARN_UNUSED __attribute__((warn_unused_result))
 
 /**
  * hiwire: A super-simple framework for converting values between C and
@@ -97,7 +98,7 @@ hiwire_incref(JsRef idval);
 /**
  * Decrease the reference count on an object.
  */
-errcode
+void
 hiwire_decref(JsRef idval);
 
 /**
@@ -191,8 +192,14 @@ JsArray_New();
 /**
  * Push a value to the end of a JavaScript array.
  */
-errcode
+errcode WARN_UNUSED
 JsArray_Push(JsRef idobj, JsRef idval);
+
+/**
+ * Same as JsArray_Push but panics on failure
+ */
+void
+JsArray_Push_unchecked(JsRef idobj, JsRef idval);
 
 /**
  * Create a new JavaScript object.
@@ -221,13 +228,13 @@ JsObject_GetString(JsRef idobj, const char* ptrname);
 /**
  * Set an object member by string.
  */
-errcode
+errcode WARN_UNUSED
 JsObject_SetString(JsRef idobj, const char* ptrname, JsRef idval);
 
 /**
  * Delete an object member by string.
  */
-errcode
+errcode WARN_UNUSED
 JsObject_DeleteString(JsRef idobj, const char* ptrname);
 
 /**
@@ -241,10 +248,10 @@ JsArray_Get(JsRef idobj, int idx);
 /**
  * Set an object member by integer.
  */
-errcode
+errcode WARN_UNUSED
 JsArray_Set(JsRef idobj, int idx, JsRef idval);
 
-errcode
+errcode WARN_UNUSED
 JsArray_Delete(JsRef idobj, int idx);
 
 /**
@@ -282,10 +289,10 @@ hiwire_call_va(JsRef idobj, ...);
 JsRef
 hiwire_call_bound(JsRef idfunc, JsRef idthis, JsRef idargs);
 
-int
+bool
 hiwire_HasMethod(JsRef obj, JsRef name);
 
-int
+bool
 hiwire_HasMethodId(JsRef obj, Js_Identifier* name);
 
 /**
@@ -471,7 +478,7 @@ hiwire_greater_than_equal(JsRef ida, JsRef idb);
 /**
  * Check if `typeof obj.next === "function"`
  */
-JsRef
+bool
 hiwire_is_iterator(JsRef idobj);
 
 /**
@@ -487,7 +494,7 @@ hiwire_next(JsRef idobj, JsRef* result);
 /**
  * Check if `typeof obj[Symbol.iterator] === "function"`
  */
-JsRef
+bool
 hiwire_is_iterable(JsRef idobj);
 
 /**
@@ -521,47 +528,35 @@ bool
 hiwire_is_typedarray(JsRef idobj);
 
 /**
- * Returns 1 if the value is a typedarray whose buffer is part of the WASM heap.
- */
-bool
-hiwire_is_on_wasm_heap(JsRef idobj);
-
-/**
- * Returns the value of `obj.byteLength`.
- *
- * There is no error checking. Caller must ensure that hiwire_is_typedarray is
- * true. If these conditions are not met, returns `0`.
- */
-int
-hiwire_get_byteLength(JsRef idobj);
-
-/**
- * Returns the value of obj.byteOffset.
- *
- * There is no error checking. Caller must ensure that hiwire_is_typedarray is
- * true and hiwire_is_on_wasm_heap is true. If these conditions are not met,
- * returns `0`.
- */
-int
-hiwire_get_byteOffset(JsRef idobj);
-
-/**
  * Copies the buffer contents of a given ArrayBuffer view or ArrayBuffer into
  * the memory at ptr.
  */
-errcode
+errcode WARN_UNUSED
 hiwire_assign_to_ptr(JsRef idobj, void* ptr);
 
 /**
  * Copies the memory at ptr into a given ArrayBuffer view or ArrayBuffer.
  */
-errcode
+errcode WARN_UNUSED
 hiwire_assign_from_ptr(JsRef idobj, void* ptr);
+
+errcode
+hiwire_write_to_file(JsRef idobj, int fd);
+
+errcode
+hiwire_read_from_file(JsRef idobj, int fd);
+
+/**
+ * Convert a buffer into a file in a copy-free manner using "canOwn" parameter.
+ * Cannot directly use the buffer anymore after using this.
+ */
+errcode
+hiwire_into_file(JsRef idobj, int fd);
 
 /**
  * Get a data type identifier for a given typedarray.
  */
-errcode
+void
 hiwire_get_buffer_info(JsRef idobj,
                        Py_ssize_t* byteLength_ptr,
                        char** format_ptr,
@@ -583,7 +578,7 @@ JsMap_New();
 /**
  * Does map.set(key, value).
  */
-errcode
+errcode WARN_UNUSED
 JsMap_Set(JsRef mapid, JsRef keyid, JsRef valueid);
 
 /**
@@ -595,7 +590,7 @@ JsSet_New();
 /**
  * Does set.add(key).
  */
-errcode
+errcode WARN_UNUSED
 JsSet_Add(JsRef mapid, JsRef keyid);
 
 #endif /* HIWIRE_H */